增加解答错误时返回具体的判题用例

+++ server/judgesvr/strategy/defaultstrategy.go
@@ -52,7 +52,7 @@ func (DefaultStrategy) ExecuteJudge(ctx *JudgeContext) (*pb.JudgeInfo, error) {
 		return judgeInfoResp, nil
 	}
 
-	if memory > memoryLimit {
+	if memory > float32(memoryLimit) {
 		judgeStatus = enum.GetJudegeInfo(enum.MemoryLimitExceeded)
 		judgeInfoResp.Message = judgeStatus
 		return judgeInfoResp, nil
